pyproject.toml
[tool.poetry]
name = "aggregate-feeds"
version = "0.1.0"
description = ""
authors = ["noreyb"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
requests = "^2.31.0"
beautifulsoup4 = "^4.12.2"
feedgenerator = "^2.1.0"
feedparser = "^6.0.10"
python-dateutil = "^2.8.2"
python-dotenv = "^1.0.0"
injector = "^0.21.0"
playwright = "^1.42.0"

[tool.poetry.group.dev.dependencies]
isort = "^5.12.0"
black = "^23.9.1"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
